fix blackboard timestamps and null values

+++ CoreEntities/Blackboard/BlackboardData.cs
@@ -4,5 +4,5 @@ public class BlackboardData<T>
 {
     public string Key { get; set; }
     public T Value { get; set; }
-    public long Timestamp { get; set; }
+    public float Timestamp { get; set; }
 }
